Move labrpc adapter to testutil package
